requirements.txt
Django==1.7.4
Pillow==2.7.0
argparse==1.2.1
coverage==3.7.1
dj-database-url==0.3.0
dj-static==0.0.6
django-cors-headers==1.0.0
django-countries==3.1.1
django-guardian==1.2.5
django-sslify==0.2.5
django-toolbelt==0.0.1
django-userena==1.4.0
djangorestframework==3.0.5
easy-thumbnails==2.2
gunicorn==19.2.1
html2text==2015.2.18
mock==1.0.1
newrelic==2.44.0.36
psycopg2==2.6
pydns==2.3.6
requests==2.5.1
sendgrid==1.3.0
six==1.9.0
smtpapi==0.2.0
static3==0.5.1
unittest2==0.8.0
wsgiref==0.1.2
